feat: create base ws server
