+++ index.ts
@@ -1,0 +1,8 @@
+import { httpServer } from "./src/http_server/index";
+import { wsServer } from "./src/ws-server/server";
+
+const HTTP_PORT = 8181;
+
+console.log(`Start static http server on the ${HTTP_PORT} port!`);
+httpServer.listen(HTTP_PORT);
+wsServer.on("listening", () => {});
